refactor(envelope): split gate owner
